requirements.txt
kafka-python==2.1.2
minio==7.1.4
python-dotenv==1.0.0